refactor(Checkbox): rename CheckBox Checkbox (change casing)
