+++ src/Checkbox/index.scss
